# MW-1031: getQuestionQttributesForEMPerformance