--- a/application/helpers/expressions/em_manager_helper.php
+++ b/application/helpers/expressions/em_manager_helper.php
@@ -8272,24 +8272,39 @@ report~numKids > 0~message~{name}, you said you are {age} and that you have {num
 #            {
 #                return $aStaticQuestionAttributesForEM[$surveyid][0][$lang][$qid];
 #            }
-
         if ($qid) {
-            $oQids = Yii::app()->db->createCommand(
-                "SELECT qid FROM {{questions}} where qid = " . intval($qid) . " and parent_qid = 0"
-            )->query()->readAll();
+            $oQids = Question::model()->findAll(
+                [
+                    'select'    => 'qid',
+                    'group'     => 'qid',
+                    'distinct'  => true,
+                    'condition' => "qid=:qid and parent_qid=0",
+                    'params'    => [':qid' => $qid]
+                ]
+            );
         } elseif ($surveyid) {
-            $oQids = Yii::app()->db->createCommand(
-                "SELECT qid FROM {{questions}} where sid = " . intval($surveyid) . " and parent_qid = 0"
-            )->query()->readAll();
+            $oQids = Question::model()->findAll(
+                [
+                    'select'    => 'qid',
+                    'group'     => 'qid',
+                    'distinct'  => true,
+                    'condition' => "sid=:sid and parent_qid=0",
+                    'params'    => [':sid' => $surveyid]
+                ]
+            );
         } else {
-            $oQids = Yii::app()->db->createCommand(
-                "SELECT qid FROM {{questions}} where parent_qid = 0"
-            )->query()->readAll();
+            $oQids = Question::model()->findAll(
+                [
+                    'select'    => 'qid',
+                    'group'     => 'qid',
+                    'distinct'  => true,
+                    'condition' => "parent_qid=0",
+                ]
+            );
         }
         $aQuestionAttributesForEM = [];
         foreach ($oQids as $oQid) {
-            $qid = $oQid["qid"];
-            $aAttributesValues = QuestionAttribute::model()->getQuestionAttributes($qid, $lang);
+            $aAttributesValues = QuestionAttribute::model()->getQuestionAttributes($oQid, $lang);
             // Change array lang to value
             foreach ($aAttributesValues as &$aAttributeValue) {
                 if (is_array($aAttributeValue)) {
@@ -8301,7 +8316,7 @@ report~numKids > 0~message~{name}, you said you are {age} and that you have {num
                     }
                 }
             }
-            $aQuestionAttributesForEM[$qid] = $aAttributesValues;
+            $aQuestionAttributesForEM[$oQid->qid] = $aAttributesValues;
         }
         EmCacheHelper::set($cacheKey, $aQuestionAttributesForEM);
         return $aQuestionAttributesForEM;