This is a performance optimization that affects survey settings tab loading.

Measurements:

Text elements	6.98s	0.49s
Privacy policy	 	6.93s	0.65s
Presentation	 	6.8s	0.45s
Participant settings	 	6.95s	0.45s
Notifications & data	 	6.87s	0.43s
Publication & access	 	6.79s	0.46s
Panel integration	 	8.23s	1.12s
Resources	 	6.89s	0.44s
Simple plugins	 	10.67s	0.46s