@Shnoulle 2024-02-22
About QuestionAttribute : a big difference between 3 and 4 are here too.

In 3 QuestionAttributedefinition was static, it's not the case (even for core part) in 4.

See [18475: Multiple event newQuestionAttributes happen](https://bugs.limesurvey.org/view.php?id=18475)

I think a good improvement can create core + plugin static function/var + update it with theme after ?