@lajosarpad 2024-02-23
@Dennis1993 @c-schmitz the change was implemented accordingly