@olleharstedt 2024-02-26
> Maybe rename variable oQids to oQuestion ?

Or get rid of Hungarian notation and just use `$question` :sweat_smile: